[API] `assign(selectWord)` throws if `selectWord` is unexpected
Not as obvious as it sounds: it means that it rejects any input
that invalidates the POSet we built, or in other words it rejects
a word if the POSet relies on the absence of it to make sense.

Other change: don't compute the set of all parent executors when there
is no executor on a node, but take the closest one.

Any input word falls in exactly one of these cases:
1. there is no executor initially defined on any path containing this node
 => API change: throws an exception, because it means there is a new branch
    in the POSet that the executor map was not supposed to cover
2. there is an executor for this word
 => no API change, return that executor
3. there is no executor on this word but one of its descendant has
 => no API change, return the executors of all children (recursively if needed)
4. there is at least an executor above this word in the POSet, and none below
 => API change: return one of the closest executors (smallest path in the graph)
5. there is at least an executor above and an executor below
 => API change: like 4. actually, return the closest executor *above*, because
    below ones are here to provide non-exhaustive specialization

Change-Id: Ic9e32ace6d3f9789adeffc345af9dffda3695977
Note: 100% test-coverage