build: Be sure to package just-processed web resources
Maven runs plugins attached to a given phase in the order of their
appearance along the POM.xml files hierarchy.

The resources processor plugin appears early, so it does not see the
resources generated by the web processor, which is attached to the
same phase but is configured later.

Configuring the resources processor _after_ the web one ensures the
processed web content is packaged.

Change-Id: I283c2297736a5a1541578b8238c605877a2cd65b