Convert inner timeouts to the appropriate HTTP status
In practice, it changes a 500 in a 504, just to not let people
think that the app is buggy when a dependency (here the DB)
takes time to answer.

Note: converting the timeout in the invoker won't be used as long
as no entry-point waits for it; it's done here "just in case" and
for consistency.

Change-Id: Id0dbeb4ad1dc81bf1ee062ab345c66d5f18e1ef7