CriteoExternalData: Remove cache
As we are observing an unexpected behavior that we were unable to
explain: the values (Maps) obtained from the cache are being
overwritten during execution. Switching to guava's cache did not solve
the issue. It is probably a race condition, but it occurs on
objects that are supposedly not shared (!).

And as the caching is not required anymore.

Change-Id: I44c9f7bbcbe5e7ff3adccb9aa4c5128d3d568988