pom.xml: (resources) Stop filtering
The only file that could matter to filter should only be
"application.conf" - and definitely NOT all groovy scripts.

As the only variable that gets filtered in "application.conf" is
"project.version", and that it is simply not used in the code,
let's just stop filtering resources entirely.

JIRA: DREDD-229
Change-Id: Iba0cadf56f3883e7da4e0246e85ea0a6414e25fd