Support Rundeck invoker in SingleTargetDispatcher mode
It's possible to configure a SingleTargetDispatcher contacting
a given job of an actual Rundeck instance, handling all deployments.

This is mainly done to ease the handling of Perpetuo when it's
installed for the first time: it allows to play with it and see
the interaction with an actual job doing something on Rundeck.

Change-Id: I60aa004f0d43d464a082cfb6cc62b3d978f628a8