perpetuo-list-editor: Stop automatically completing
When a single element is filtered by the input, do not automatically
use it. While the idea seems good in terms of efficiency, it is
unconventional, and people complained about it (e.g. when doing a typo
leading to only one element, this one gets today immediately
completed)

Let's just keep a single element in the drop down, and let the user
decide what she wants to do with it.

Change-Id: Ie7cf1bb1f9cc4c62cb11b3871624aa8bb6f0d0cb