Cosmetics (Schema)
About the disappearance of mapValues, which might seem as not being
cosmetics, it is: the provided lambda (converting a TargetStatus)
is called on another line, to be consistent to how and where
the ExecutionTraces are converted. But the evaluation of the lambda
already took place there, because mapValues only registers the lambda
to be applied each time a get() is invoked for instance.

Change-Id: Iec68c79d7027a1eb6ea966f2be3251585cb20f68