Listeners: (onCreatingDeploymentRequests) Introduce event-preceeding-action
The intent is to allow listeners to veto an action before it happens,
by raising an exception (which will be ultimately transmitted to the
front-end). This will notably make it possible to move out the
handling of the "old workflow" back to Criteo's specifics, and enable
the support of black-list based veto (before the start of a deployment
operation).

Change-Id: Iafd98fa7ae329de20c31311fbd8a3c5fde797557