Server: (config) Support Logback's configuration file selection
Introducing a ServerConfigurator trait to ensure that the
configuration is read and applied to Logback before anything else
happens.

Logback's documentation specifies that the selection of the
configuration file can be made by setting a dedicated system property,
but that it must be set before any logger instance is created. Setting
the system property in the very first lines of the Server class
doesn't work, because its base trait (Finatra's HttpServer) logs
things (=> already instantiate loggers). So the property must be set
_before_ the initialization of the base trait; this is accomplished
here by introducing a new trait ServerConfigurator mixed _before_
HttpServer.

Change-Id: I60eb11107466240a20b493d6d6d5f4d44b804e8a