ExecutionHttpTrigger: (ctor) Compute "ssl" on demand
Working around the current issue on valuation order ("port" is set at
the moment after the computation of "ssl").

Note anyway that the thing needs to be reworked, as the root problem
is not globaly fixed: if another use of "port" or "host" is made at
trait construction time, no error at compilation is raised despite
them not being defined.

Change-Id: I167988847d2bee2cf22b0859ef201064aae784bb