Crankshaft: (tryUpdateExecutionTrace) No transaction
The transaction favors a race condition where two simultaneous updates
close two open execution traces but none of them sees that no execution
trace remains open after that, so none of the two requests closes the
operation trace.

The goal of the transaction was precisely to avoid having open operations
without any open executions, but on a contrary, it favored this situation!

Change-Id: Ic6ec9647a6c0a5ba528d59c3f595d9efc9d4d17f