deployment-request-table: Prefer a custom helper to manipulate location
The current use of app-location is problematic due to the changing of
the execution order of properties settings and related event
emissions.

Indeed, when the app is first displayed, the elements are created and
properties callbacks are invoked - and this.queryParams is first set,
then onActiveChanged is. But when navigating the app - getting back to
the deployment request table - the execution order is reversed:
onActiveChanged is invoked before the queryParams are known,
(triggering the observed bug). Add to that the fact that the selected
product should be set once the list of products is field, which is
done when the asynchronous fetching of the products is resolved.

Solving this with app-location would be cumbersome and make the code
more complicated than needed by a simple, synchronous - but alas,
custom - query parameters extractor (tested on Chrome and Firefox).

JIRA: DREDD-1000
Change-Id: Ib627c6c0a974313e7cda94f1bbde31128892ab7a