FineGrainedPermissions: (ProductRule) Cosmetics: class instead of case class
The primary purpose of case classes is pattern matching.
Pattern matching doesn't make sense if the attributes
are not comparable, like Pattern here. So let's not give
a false impression (the support of pattern matching)
about a type (ProductRule) by using the wrong way of
defining it (as a case class).

Change-Id: Ia8b7e6289f36c7e3562eecb5431751bdd7c414f1