Engine, Crankshaft: Prefer raising explicit/concrete exceptions on unexpected state
Over raising Conflict or UnavailableAction with a text description.

Rationale: this pushes back the serialization (which in these cases is
essentially a human-readable long description of the type) to the
catcher, which may or not choose to do it. In tests notably, it
enables to assess a situation by looking at the type of an exception
emitted, instead of working on text messages.

Change-Id: Icb6c6332e37e797bd53d128d4646e9bb725c4c38