TargetAtomSet: Expose `subset` and `superset` instead of `items` and `isExact`
So that we know which atoms are expected to be impacted and which
are covered as a result of an "inexact" resolution but may or may
not be impacted.

It fixes the issue of all atoms being set as notDone before the
deployment start, possibly making some deployments wrongly
considered as failed.

Change-Id: Id11f6053740bc9472c32a1e53803c8b28861074c