deployment-request-table: onActiveChanged: Resume refresh when all set
Refreshing takes into account various parameters (the selected product
name if any, the current page, the page size), which can be set by the
query part of the URL. Hence, the URL must be interpreted - and said
parameters set - before resuming.

Besides, the selected product name can only be set after the list of
products has been retrieved - as the former actually selects an item
in the latter.

Thus, the flow is:
1. retrieve the list of products;
2. interpret the URL, to possibly select a product name, a page, and
   a page size;
3. resume refresh

Change-Id: I85291d7923300f7e0163f2d58c3be7701ab717a4